requirements.txt
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-debug-toolbar==1.9.1
django-nose==1.4.5
factory-boy==2.9.2
Faker==0.8.7
idna==2.6
nose==1.3.7
oauthlib==2.0.6
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.7.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.1
urllib3==1.22
